Split printPage.js
Split all the functions from printPage into their own files.
Either \[function\]Content.js or \[function\]Code.js
