+++ printModules/pageContent/pageLoggedIn/loginSuccessCode.js
@@ -1,0 +1,12 @@
+import buildWheel from "./modules/wheel.js";
+
+export default function loginSuccessCode() {
+    buildWheel();
+    let betBtn = document.getElementById("betBtn");
+    //let wagerInput = document.getElementById("wagerInput");
+    //let betInput = document.getElementById("betInput");
+    //let wagerInputValue = wagerInput.value;
+    //let betInputValue = betInput.value;
+    //betBtn.addEventListener("click", betBtnEl);
+    //betBtnEl(wagerInput, betInput);
+}
