+++ printModules/pageContent/pageLoggedOut/pageUnknown/failedLoginContent.js
@@ -1,6 +1,5 @@
-export default function failedText(){
-let failText= 
-contentDiv.innerHTML = `
+export default function failedLoginContent(){
+let failText =  `
 <div id="errorMessage">
 <h2>Wrong Credentials</h2>
 <p>Invalid username or password. <br>
